Moved rs256.go and hs256.go into a separated algorithms package
