+++ token/algorithm/hs256.go
@@ -1,4 +1,4 @@
-package token
+package algorithm
 
 import (
 	"github.com/golang-jwt/jwt/v5"
