+++ token/algorithm/rs256.go
@@ -1,11 +1,16 @@
-package token
+package algorithm
 
 import (
 	"fmt"
 	"github.com/golang-jwt/jwt/v5"
+	credstackError "github.com/stevezaluk/credstack-lib/errors"
 	"github.com/stevezaluk/credstack-lib/key"
+
 	keyModel "github.com/stevezaluk/credstack-lib/proto/key"
 )
+
+// ErrFailedToSignToken - An error that gets wrapped when jwt.Token.SignedString returns an error
+var ErrFailedToSignToken = credstackError.NewError(500, "ERR_FAILED_TO_SIGN", "token: Failed to sign token due to an internal error")
 
 /*
 GenerateRS256 - Generates arbitrary RS256 tokens with the claims that are passed as an argument to this function. This
